undo get asset icons consolidation
